Segregate Projects In Respective Folder Levels
